move ads to bottom deactive keep awake on settings page
